+++ Hybrid/vtkPointWidget.h
@@ -86,11 +86,11 @@ public:
   // Description:
   // Methods that satisfy the superclass' API.
   virtual void SetEnabled(int);
-  virtual void PlaceWidget(float bounds[6]);
+  virtual void PlaceWidget(double bounds[6]);
   void PlaceWidget()
     {this->Superclass::PlaceWidget();}
-  void PlaceWidget(float xmin, float xmax, float ymin, float ymax, 
-                   float zmin, float zmax)
+  void PlaceWidget(double xmin, double xmax, double ymin, double ymax, 
+                   double zmin, double zmax)
     {this->Superclass::PlaceWidget(xmin,xmax,ymin,ymax,zmin,zmax);}
 
   // Description:
@@ -102,13 +102,13 @@ public:
   // Set/Get the position of the point. Note that if the position is set
   // outside of the bounding box, it will be clamped to the boundary of
   // the bounding box.
-  void SetPosition(float x, float y, float z)
+  void SetPosition(double x, double y, double z)
     {this->Cursor3D->SetFocalPoint(x,y,z);}
-  void SetPosition(float x[3])
+  void SetPosition(double x[3])
     {this->SetPosition(x[0],x[1],x[2]);}
-  float* GetPosition() 
+  double* GetPosition() 
     {return this->Cursor3D->GetFocalPoint();}
-  void GetPosition(float xyz[3]) 
+  void GetPosition(double xyz[3]) 
     {this->Cursor3D->GetFocalPoint(xyz);}
 
   // Description:
@@ -197,8 +197,8 @@ public:
   // motion vector is used to control the constrained sliding action. Note the
   // size is specified as a fraction of the length of the diagonal of the 
   // point widget's bounding box.
-  vtkSetClampMacro(HotSpotSize,float,0.0,1.0);
-  vtkGetMacro(HotSpotSize,float);
+  vtkSetClampMacro(HotSpotSize,double,0.0,1.0);
+  vtkGetMacro(HotSpotSize,double);
   
 protected:
   vtkPointWidget();
@@ -256,7 +256,7 @@ protected:
   void CreateDefaultProperties();
   
   // The size of the hot spot.
-  float HotSpotSize;
+  double HotSpotSize;
   int DetermineConstraintAxis(int constraint, double *x);
   int WaitingForMotion;
   int WaitCount;
